Edit javadocs
