+++ server/src/main/java/io/javaclasses/filehub/storage/UserRecord.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
- * A record about a particular user. Should be saved in a UserStorage.
+ * A {@link StorageRecord} about a particular user. Should be saved in a {@link UserStorage}.
  */
 @Immutable
 public final class UserRecord implements StorageRecord<UserId> {
